hidden buffer is not a hidden rocketchat room
No longer tie weechat's concept of hidden to that of rocketchat.
Instead, map a closed buffer to a hidden room which is a better match
as weechat buffers are more for window management.

This required revisiting how a disconnect from the server was handled
and cleaning up of how buffers were closed.  In the end, a new server
state was introduced, DISCONNECTING, which allows differentiating
between a buffer close due to a disconnect vs a regular close requested
by the user.