add minimal support for quoted messages
This could be better, but it shows the last message for context and puts
the whole quoted block, including the quoter's nick, in italics.  That
stands out a bit and at least isn't top-posting like the web UI.

In an ideal world, we'd display the 2nd level quote, if any, that's
already included in the streamed message.  Then we'd properly indent,
taking care to split lines that are longer than the chat window
appropriately, and potentially use colors for each level of quoting.